+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fixed
 - Throwing a non-error from tests breaks various reporters
+- Replaced deprecated assert.strictEqual with assert.strictEqual
 
 ### Added
 - Discover tripitaka badge